+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## unreleased
+- Update Android default Version to 1.1.4 --> Enabled SDK 30 support
+- Update iOS Pod to 4.4.0 --> [Changelog](https://github.com/Yummypets/YPImagePicker/releases)
+- **BREAKING**: iOS know works with a file url and only returns isBase64 if `asBase64: true` was set
+- Function `cleanup` added for ios
+
+
 ## 1.0.1
 - I changed my GitHub Username, so i need to adjust stuff
 - Typing fixed (closes [#4](https://github.com/EinfachHans/cordova-plugin-advanced-imagepicker/issues/4))